Refactor use of ssh attributes and overall code.
This commit refactors (for a third time!) the use of ssh configuration
attributes. Specifically, it aligns with the Vagrant 1.2 direction
where the config.ssh.* attributes override everything including the
provider.

Additionally, the entire codebase was cleaned up for consistent style
and internationalization messages in preparation for a 0.1.0 release.

Updated documentation will come shortly.