Change default image and size
Ubuntu 14.04 is no longer supported by Canonical so this commit updates the
default to Ubuntu 18.04.

The `512mb` slug is deprecated. This commit changes the default slug to `s-1vcpu-1gb`. This addresses issue #280.

This commit also updates the README file to use the new image names and
slugs, as well as specifies the `rsync` method for syncing in the
example code, which is required on Vagrant 2. This addresses issue #277.